requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-import-export==2.2.0
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0
